+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix issue [#312](https://github.com/intersystems/language-server/issues/312): Fix routine existence diagnostics for routines that only exist in OBJ form
 - Fix issue [#313](https://github.com/intersystems/language-server/issues/313): Add Diagnostic when `ROUTINE` header is missing
 - Fix issue [#314](https://github.com/intersystems/language-server/issues/314): Suggest boolean class keywords for completion after typing `Not`
+- Fix issue [#316](https://github.com/intersystems/language-server/issues/316): Show class description when hovering over a typed variable
 - Fix issue [#317](https://github.com/intersystems/language-server/issues/317): Add `DocumentLink` for `##class()` in class description comments
 - Parser changes:
   - DP-430347: Track variables in routine procedure blocks